refactor: move SelectionRangeAdapter to new file
