+++ src/language/common/SelectionRangeAdapter.ts
@@ -1,0 +1,43 @@
+import * as lsTypes from 'vscode-languageserver-types';
+import { languages, editor, Position, CancellationToken } from '../../fillers/monaco-editor-core';
+import { fromPosition, toRange } from './CompletionAdapter';
+import { WorkerAccessor } from './lspLanguageFeatures';
+
+export interface ILanguageWorkerWithSelectionRanges {
+	getSelectionRanges(uri: string, positions: lsTypes.Position[]): Promise<lsTypes.SelectionRange[]>;
+}
+
+export class SelectionRangeAdapter<T extends ILanguageWorkerWithSelectionRanges>
+	implements languages.SelectionRangeProvider
+{
+	constructor(private _worker: WorkerAccessor<T>) {}
+
+	public provideSelectionRanges(
+		model: editor.IReadOnlyModel,
+		positions: Position[],
+		token: CancellationToken
+	): Promise<languages.SelectionRange[][] | undefined> {
+		const resource = model.uri;
+
+		return this._worker(resource)
+			.then((worker) =>
+				worker.getSelectionRanges(
+					resource.toString(),
+					positions.map<lsTypes.Position>(fromPosition)
+				)
+			)
+			.then((selectionRanges) => {
+				if (!selectionRanges) {
+					return;
+				}
+				return selectionRanges.map((selectionRange: lsTypes.SelectionRange | undefined) => {
+					const result: languages.SelectionRange[] = [];
+					while (selectionRange) {
+						result.push({ range: toRange(selectionRange.range) });
+						selectionRange = selectionRange.parent;
+					}
+					return result;
+				});
+			});
+	}
+}
